+++ backend/src/main/java/com/cursoIntegrador/lePettiteCoffe/LePettiteCoffeApplication.java
@@ -2,10 +2,8 @@ package com.cursoIntegrador.lePettiteCoffe;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 
-@SpringBootApplication // (exclude = { DataSourceAutoConfiguration.class }) // LA ANOTACION ES TEMPORAL
-						// HASTA TENER BD
+@SpringBootApplication
 public class LePettiteCoffeApplication {
 
 	public static void main(String[] args) {
