+++ backend/src/main/resources/application.properties
@@ -1,9 +1,8 @@
 spring.application.name=lePettiteCoffe
 
-### Comentar si se quiere apagar el datasource
 spring.datasource.url=jdbc:postgresql://localhost:5432/LePettiteCoffe
 spring.datasource.username=postgres
-spring.datasource.password=password
+spring.datasource.password=contraseña
 spring.datasource.driver-class-name=org.postgresql.Driver
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
